common/fakes_segment/requirements.txt
torch==2.2.0
transformers==4.37.2
pillow==10.2.0
matplotlib==3.7.4
